objection/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
commonmark==0.9.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gevent==24.2.1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-translate==2.0.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.66.0
grpcio-status==1.48.2
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lxml==4.9.4
marisa-trie==1.2.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
ml-dtypes==0.4.0
Morfessor==2.0.6
moviepy==1.0.3
murmurhash==1.0.10
namex==0.0.8
nltk==3.9.1
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
Pillow==9.5.0
polyglot==16.7.4
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
proglog==0.1.10
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycld2==0.41
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyICU==2.13.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==12.6.0
rsa==4.9
safetensors==0.4.4
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spaw==0.2
srsly==2.4.8
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-cpu==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
textblob==0.17.1
tf_keras==2.17.0
thinc==8.2.5
tinydb==4.8.0
tokenizers==0.19.1
toml==0.10.2
tqdm==4.66.5
transformers==4.44.2
typer==0.12.4
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
zope.event==5.0
zope.interface==7.0.1
